introducted some styling with material ui
